requirements.txt
streamlit>=1.28.0
ultralytics>=8.3.217
opencv-python-headless>=4.12.0.88
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0
scikit-learn>=1.3.0
psutil>=5.9.0
